RAG/rag_lab/pyproject.toml
[project]
name = "rag-lab"
version = "0.1.0"
description = "RAG component lab bench — swap components, measure deltas, build intuition"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0.0",
    "chromadb>=0.5.0",
    "sentence-transformers>=3.0.0",
    "ragas>=0.2.0",
    "pypdf>=4.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "numpy>=1.26.0",
    "datasets>=2.20.0",
    "langchain-openai>=0.3.0",
    "langchain-text-splitters>=0.3.0",
    "langchain-experimental>=0.4.1",
    "langchain-huggingface>=0.1.0",
    "lxml>=5.0.0",
    "bs4>=0.0.2",
    "llama-index-core>=0.10.0",
    "llama-index-embeddings-huggingface>=0.2.0",
    "tree-sitter>=0.21.0",
    "tree-sitter-language-pack>=0.7.0",
]
